clean index.tsx -> SendEmail.tsx component created
